requirements.txt
folium==0.14.0
geopandas==0.14.0
matplotlib==3.8.0
numpy==1.26.1
osmnx==1.7.1
pandas==2.1.2
Pillow==10.1.0
Pillow==10.1.0
Shapely==2.0.2
